toggl/operations#1192 ping toggl api on status check
